[CSL 1413] .net sdk clean up (#8)
* Resolve lint warnings

* Resolve lint warnings

* Resolve warnings

* Add documentation to request classes

* Add config class

* Update docs to modules

* Remove unused json

* Add config class and readme

* Edit readme and config class

+++ src/Constructorio_NET/models/Recommendations/RecommendationsResponseInner.cs
@@ -7,16 +7,15 @@ using Newtonsoft.Json;
  */
 namespace Constructorio_NET.Models
 {
-    public class RecommendationsResponseInner {
-
+    public class RecommendationsResponseInner
+    {
         [JsonProperty("results")]
         public List<Result> Results { get; set; }
 
         [JsonProperty("total_num_results")]
-        public Int16 TotalNumberOfResults { get; set; }
+        public int TotalNumberOfResults { get; set; }
 
         [JsonProperty("pod")]
         public ResultPod Pod { get; set; }
-
     }
 }
